Applied Spatial Audio (Not Left-Right Yet)

+++ Order_227/Motor2D/Geometry.cpp
@@ -15,7 +15,7 @@ fVec2 GetVector2(fPoint A, fPoint B)
 float GetDistance(iPoint A, iPoint B)
 {
 	iVec2 vec = GetVector2(A, B);
-	return (float)vec.GetMagnitude();
+	return vec.GetMagnitude();
 }
 
 float GetDistance(fPoint A, fPoint B)
